+++ apps/tickets/infrastructure/services/kilometrage_repository.py
@@ -3,6 +3,7 @@
 from __future__ import annotations
 
 from datetime import date
+from decimal import Decimal
 
 from django.db.models import Sum
 
@@ -12,7 +13,9 @@ from apps.tickets.infrastructure.models import KilometrageRecordModel
 class KilometrageRepository:
     """Provide kilometrage lookups backed by the database."""
 
-    def get_km_at_or_before(self, unit_number: str, target_date: date) -> int | None:
+    def get_km_at_or_before(
+        self, unit_number: str, target_date: date
+    ) -> Decimal | None:
         """Return kilometer value at or before a given date.
 
         Args:
@@ -34,7 +37,7 @@ class KilometrageRepository:
         )
         return record.km_value if record else None
 
-    def get_km_since(self, unit_number: str, from_date: date) -> int | None:
+    def get_km_since(self, unit_number: str, from_date: date) -> Decimal | None:
         """Return total kilometers since a given date.
 
         Sums all km values from the first record on or after from_date
@@ -55,7 +58,7 @@ class KilometrageRepository:
         ).aggregate(total=Sum("km_value"))
         return result["total"] if result["total"] is not None else None
 
-    def get_latest_km(self, unit_number: str) -> int | None:
+    def get_latest_km(self, unit_number: str) -> Decimal | None:
         """Return latest kilometer value for a unit."""
 
         unit_key = unit_number.strip().upper()
